Add support for multiple data types (remainder of #196) (#203)
Addresses #95 and #111.
Follow-up to #198, #199, #201

Trying again, since #130 failed. This time, I made the Model class to be polymorphic. This way, the amount of pointer indirection is minimized.

Summary: Model is an opaque container that wraps the polymorphic handle ModelImpl<ThresholdType, LeafOutputType>. The handle in turn stores the list of trees Tree<ThresholdType, LeafOutputType>. To unbox the Model container and obtain ModelImpl<ThresholdType, LeafOutputType>, use Model::Dispatch(<lambda expression>).

Also, upgrade to C++14 to access the generic lambda feature, which proved to be very useful in the dispatching logic for the polymorphic Model class.

* Turn the Model and Tree classes into template classes
* Revise the string templates so that correct data types are used in the generated C code
* Rewrite the model builder class
* Revise the zero-copy serializer
* Create an abstract matrix class that supports multiple data types (float32, float64 for now).
* Move the DMatrix class to the runtime.
* Extend the DMatrix class so that it can hold float32 and float64.
* Redesign the C runtime API using the DMatrix class.
* Ensure accuracy of scikit-learn models. To achieve the best results, use float32 for the input matrix and float64 for the split thresholds and leaf outputs.
* Revise the JVM runtime.

+++ src/compiler/common/format_util.h
@@ -66,7 +66,7 @@ class ArrayFormatter {
    */
   ArrayFormatter(size_t text_width, size_t indent, char delimiter = ',')
     : oss_(), text_width_(text_width), indent_(indent), delimiter_(delimiter),
-      default_precision_(oss_.precision()), line_length_(indent),
+      default_precision_(static_cast<int>(oss_.precision())), line_length_(indent),
       is_empty_(true) {}
 
   /*!
